moved FieldAdapter back to adapters.py
