+++ tests/test_adapters.py
@@ -1,211 +1,19 @@
 from typing import get_args
-from pydantic import TypeAdapter, Discriminator, Tag
+from pydantic import Discriminator, Tag
 from pymarc import Field as MarcField
 import pytest
 from record_validator.adapters import (
-    get_adapter,
-    get_record_type,
-    pattern_match,
     tag_discriminator,
     AuxOtherFields,
     MonographFields,
     OtherFields,
+    FieldAdapter,
 )
 from record_validator.field_models import (
     AuxBibCallNo,
     MonographDataField,
     OtherDataField,
 )
-
-
-def test_get_adapter(stub_record):
-    record_type, adapter = get_adapter(stub_record.fields)
-    assert isinstance(record_type, str)
-    assert isinstance(adapter, TypeAdapter)
-
-
-def test_get_adapter_monograph_evp(stub_record):
-    record_type, adapter = get_adapter(stub_record.fields)
-    schema = adapter.json_schema(by_alias=True)
-    assert record_type == "evp_monograph"
-    assert isinstance(adapter, TypeAdapter)
-    assert sorted([i for i in schema["$defs"].keys()]) == [
-        "BibCallNo",
-        "BibVendorCode",
-        "ControlField001",
-        "ControlField003",
-        "ControlField005",
-        "ControlField006",
-        "ControlField007",
-        "ControlField008",
-        "InvoiceField",
-        "ItemField",
-        "LCClass",
-        "LibraryField",
-        "MonographDataField",
-        "OrderField",
-    ]
-
-
-def test_get_adapter_monograph_no_vendor(stub_record):
-    stub_record.remove_fields("901")
-    stub_record["949"].delete_subfield("v")
-    record_type, adapter = get_adapter(stub_record.fields)
-    schema = adapter.json_schema(by_alias=True)
-    assert record_type == "monograph"
-    assert isinstance(adapter, TypeAdapter)
-    assert sorted([i for i in schema["$defs"].keys()]) == [
-        "BibCallNo",
-        "BibVendorCode",
-        "ControlField001",
-        "ControlField003",
-        "ControlField005",
-        "ControlField006",
-        "ControlField007",
-        "ControlField008",
-        "InvoiceField",
-        "ItemField",
-        "LCClass",
-        "LibraryField",
-        "MonographDataField",
-        "OrderField",
-    ]
-
-
-def test_get_adapter_other_auxam(stub_aux_other_record):
-    record_type, adapter = get_adapter(stub_aux_other_record.fields)
-    schema = adapter.json_schema(by_alias=True)
-    assert record_type == "auxam_other"
-    assert isinstance(adapter, TypeAdapter)
-    assert sorted([i for i in schema["$defs"].keys()]) == [
-        "AuxBibCallNo",
-        "BibVendorCode",
-        "ControlField001",
-        "ControlField003",
-        "ControlField005",
-        "ControlField006",
-        "ControlField007",
-        "ControlField008",
-        "InvoiceField",
-        "LCClass",
-        "LibraryField",
-        "OrderField",
-        "OtherDataField",
-    ]
-
-
-def test_get_adapter_other_evp(stub_pamphlet_record):
-    record_type, adapter = get_adapter(stub_pamphlet_record.fields)
-    schema = adapter.json_schema(by_alias=True)
-    assert record_type == "evp_other"
-    assert isinstance(adapter, TypeAdapter)
-    assert sorted([i for i in schema["$defs"].keys()]) == [
-        "BibVendorCode",
-        "ControlField001",
-        "ControlField003",
-        "ControlField005",
-        "ControlField006",
-        "ControlField007",
-        "ControlField008",
-        "InvoiceField",
-        "LCClass",
-        "LibraryField",
-        "OrderField",
-        "OtherDataField",
-    ]
-
-
-def test_get_adapter_other_no_vendor(stub_pamphlet_record):
-    stub_pamphlet_record.remove_fields("901")
-    record_type, adapter = get_adapter(stub_pamphlet_record.fields)
-    schema = adapter.json_schema(by_alias=True)
-    assert record_type == "other"
-    assert isinstance(adapter, TypeAdapter)
-    assert sorted([i for i in schema["$defs"].keys()]) == [
-        "BibVendorCode",
-        "ControlField001",
-        "ControlField003",
-        "ControlField005",
-        "ControlField006",
-        "ControlField007",
-        "ControlField008",
-        "InvoiceField",
-        "LCClass",
-        "LibraryField",
-        "OrderField",
-        "OtherDataField",
-    ]
-
-
-def test_get_record_type_monograph(stub_record):
-    assert get_record_type(fields=stub_record.fields, vendor_code="EVP") == "monograph"
-
-
-def test_get_record_type_dance(stub_dance_record):
-    assert (
-        get_record_type(fields=stub_dance_record.fields, vendor_code="EVP") == "other"
-    )
-
-
-def test_get_record_type_pamphlet(stub_pamphlet_record):
-    assert (
-        get_record_type(fields=stub_pamphlet_record.fields, vendor_code="EVP")
-        == "other"
-    )
-
-
-def test_get_record_type_catalogue(stub_catalogue_record):
-    assert (
-        get_record_type(fields=stub_catalogue_record.fields, vendor_code="EVP")
-        == "other"
-    )
-
-
-def test_get_record_type_multivol(stub_multivol_record):
-    assert (
-        get_record_type(fields=stub_multivol_record.fields, vendor_code="EVP")
-        == "other"
-    )
-
-
-def test_get_record_type_auxam_other(stub_aux_other_record):
-    assert (
-        get_record_type(fields=stub_aux_other_record.fields, vendor_code="AUXAM")
-        == "other"
-    )
-
-
-def test_get_record_type_bad_input():
-    assert get_record_type(fields=["foo"], vendor_code="AUXAM") == "other"
-
-
-def test_get_record_type_dict(stub_record):
-    record_dict = stub_record.as_dict()
-    assert (
-        get_record_type(fields=record_dict["fields"], vendor_code="EVP") == "monograph"
-    )
-
-
-@pytest.mark.parametrize(
-    "subfields, pattern, expected",
-    [
-        (["2 volumes :"], "multivol", True),
-        (["2 volumes :"], "pamphlet", False),
-        (["5 pages :", "illustrations"], "pamphlet", True),
-        (["5 pages :", "illustrations"], "catalogues", False),
-        (["Catalogues Raisonnes"], "catalogues", True),
-        (["Catalogues Raisonnés"], "catalogues", True),
-        (["Catalogue Raisonnés"], "catalogues", True),
-        (["foo"], "catalogues", False),
-        (["ReCAP 24-"], "aux_call_no", True),
-        (["ReCAP 25-"], "aux_call_no", True),
-        (["ReCAP 24-100000"], "aux_call_no", False),
-        (["foo"], "aux_call_no", False),
-        (None, "foo", False),
-    ],
-)
-def test_pattern_match(subfields, pattern, expected):
-    assert pattern_match(subfields, pattern) == expected
 
 
 @pytest.mark.parametrize(
@@ -273,3 +81,27 @@ def test_OtherFields():
     assert (OtherDataField, Tag("data_field")) in union_args
     assert (OtherDataField, Tag("852")) in union_args
     assert (OtherDataField, Tag("949")) in union_args
+
+
+def test_FieldAdapter():
+    schema = FieldAdapter.json_schema(by_alias=True)
+    assert sorted([i for i in schema["$defs"].keys()]) == sorted(
+        [
+            "AuxBibCallNo",
+            "BibCallNo",
+            "BibVendorCode",
+            "ControlField001",
+            "ControlField003",
+            "ControlField005",
+            "ControlField006",
+            "ControlField007",
+            "ControlField008",
+            "InvoiceField",
+            "ItemField",
+            "LCClass",
+            "LibraryField",
+            "MonographDataField",
+            "OrderField",
+            "OtherDataField",
+        ]
+    )
