+++ tests/test_field_models.py
@@ -18,7 +18,6 @@ from record_validator.field_models import (
     MonographDataField,
     OrderField,
     OtherDataField,
-    FieldAdapter,
 )
 
 
@@ -1649,27 +1648,3 @@ def test_OtherDataField_invalid_tag_literal(tag):
     error_types = [error["type"] for error in e.value.errors()]
     assert len(e.value.errors()) == 1
     assert error_types == ["string_pattern_mismatch"]
-
-
-def test_FieldAdapter():
-    schema = FieldAdapter.json_schema(by_alias=True)
-    assert sorted([i for i in schema["$defs"].keys()]) == sorted(
-        [
-            "AuxBibCallNo",
-            "BibCallNo",
-            "BibVendorCode",
-            "ControlField001",
-            "ControlField003",
-            "ControlField005",
-            "ControlField006",
-            "ControlField007",
-            "ControlField008",
-            "InvoiceField",
-            "ItemField",
-            "LCClass",
-            "LibraryField",
-            "MonographDataField",
-            "OrderField",
-            "OtherDataField",
-        ]
-    )
